+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- [Memory cards](https://github.com/Radiokot/photoprism-android-client/wiki/Memories-extension)
+  no longer move around immediately after being seen, so they behave more like Instagram stories
+
 ## [1.35.0] - 2025-02-04
 
 ### Added